under_planet_final/under_planet_grace/requirements.txt
anyio==3.7.0
apturl==0.5.2
asgiref==3.6.0
attrs==21.2.0
Authlib==1.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.26.96
botocore==1.29.96
Brlapi==0.8.3
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
cryptography==39.0.0
cupshelpers==1.0
dateparser==1.1.8
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20200713
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crontab==0.7.1
django-import-export==3.1.0
django-js-asset==2.0.0
django-sslserver==0.22
django-storages==1.13.2
djangorestframework==3.14.0
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
ecdsa==0.18.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
humanfriendly==10.0
idna==3.4
imageio==2.25.1
img2pdf==0.4.4
importlib-metadata==4.6.4
imutils==0.5.4
jeepney==0.7.1
jmespath==1.0.1
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.0
notify2==0.3
numpy==1.24.2
oauthlib==3.2.2
ocrmypdf==14.0.3
odfpy==1.4.1
opencv-python==4.7.0.68
openpyxl==3.1.1
packaging==23.0
paramiko==2.9.3
pdf2image==1.16.3
pdfkit==1.0.0
pdfminer.six==20221105
pexpect==4.8.0
pikepdf==7.0.0
Pillow==9.4.0
platformdirs==2.5.1
pluggy==1.0.0
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyactiveresource==2.2.2
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.8
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pypng==0.20220715.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-jose==3.3.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.7.1
pyuploadcare==4.0.0
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0
qrcode==7.4.2
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3cmd==2.3.0
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.10.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
ShopifyAPI==12.3.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
systemd-python==234
tablib==3.3.0
tesserocr==2.5.2
texttable==1.6.4
tifffile==2023.2.3
tqdm==4.64.1
typing_extensions==4.5.0
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uploadcare==0.1.1
urllib3==1.26.15
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
websocket-client==1.2.3
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
